+++ src/settingsdialog.cpp
@@ -44,13 +44,14 @@ SettingsDialog::SettingsDialog(QWidget *parent) :
 
     // Set languages data
     ui->languageComboBox->setItemData(0, QLocale::AnyLanguage);
-    ui->languageComboBox->setItemData(1, QLocale::German);
+    ui->languageComboBox->setItemData(1, QLocale::Chinese);
     ui->languageComboBox->setItemData(2, QLocale::English);
     ui->languageComboBox->setItemData(3, QLocale::Finnish);
-    ui->languageComboBox->setItemData(4, QLocale::Portuguese);
-    ui->languageComboBox->setItemData(5, QLocale::Spanish);
+    ui->languageComboBox->setItemData(4, QLocale::German);
+    ui->languageComboBox->setItemData(5, QLocale::Portuguese);
     ui->languageComboBox->setItemData(6, QLocale::Russian);
-    ui->languageComboBox->setItemData(7, QLocale::Chinese);
+    ui->languageComboBox->setItemData(7, QLocale::Spanish);
+    ui->languageComboBox->setItemData(8, QLocale::Turkish);
 
     loadSettings();
 }