+++ CHANGELOG.md
@@ -6,7 +6,9 @@ All notable changes to this project will be documented in this file. This projec
 
 [Full Changelog](https://github.com/Shatur95/optimus-manager-qt/compare/1.2.6...HEAD)
 
-_No unreleased changes yet_
+**Added**
+
+-   Turkish translation.
 
 ## [1.2.6](https://github.com/Shatur95/optimus-manager-qt/tree/1.2.6) (2019-07-21)